% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{enplot}
\alias{enplot}
\title{Score Plot for Proficiency Testing}
\usage{
enplot(x,title,xlabel)
}
\arguments{
\item{x}{ a numeric vector of data values. Data values must be En score.}
\item{title}{ overall title for the plot.}
\item{xlabel}{ a label for the x axis.}
}
\details{
ISO 13528:2015 provides detailed descriptions of statistical methods for proficiency testing providers to use to design proficiency testing schemes and to analyse the data obtained from those schemes. Some of these statistical methods are called  performance statistics and the most widely used are zscore, z'score, zeta score and En score.
}
\description{
This function provides En score plot.
}
\author{
Luiz Henrique Leal \email{lhleal@inmetro.gov.br}
}
\references{
ISO 13528:2015(E) Statistical methods for use in proficiency testing by interlaboratory comparison. Second edition.
}
\seealso{
  \code{\link{zplot}}.
}
\examples{
 enscore=seq(-4,4)
 names(enscore)=paste("L",1:length(enscore),sep="")
 par(font=2,cex.axis=1.3,cex.main=1.7,cex.lab=1.3)
 enplot(x=enscore,title="NOx",xlabel="Laboratory")

}
